test(gui): restructured e2e tests to ease reuse w/ other clients
+ extended dependency structure to enable parallelism where possible
