+++ frontend/tests/e2e_tests/integration/01-basic/01-basic.spec.ts
@@ -13,12 +13,16 @@
 //    limitations under the License.
 import { expect } from '@playwright/test';
 
-import test from '../fixtures/fixtures.ts';
-import { emptyStorageState } from '../utils/constants.ts';
+import test from '../../fixtures/fixtures';
+import { emptyStorageState } from '../../utils/constants';
 
 test.use({ storageState: { ...emptyStorageState } });
 
 test.describe('Basic functionality checks', () => {
+  let navbar;
+  test.beforeEach(async ({ page }) => {
+    navbar = page.locator('.leftFixed.leftNav');
+  });
   test.describe('basic window checks', () => {
     test('get the global window object', async ({ baseUrl, context, page }) => {
       page = await context.newPage();
@@ -35,6 +39,21 @@ test.describe('Basic functionality checks', () => {
       page = await context.newPage();
       await page.goto(`${baseUrl}ui/`);
       await expect(page).toHaveTitle(/Mender/i);
+    });
+  });
+  test.describe('Overall layout and structure', () => {
+    test('shows the left navigation', async () => {
+      await expect(navbar.getByRole('link', { name: /Dashboard/i })).toBeVisible();
+      await expect(navbar.getByRole('link', { name: /Devices/i })).toBeVisible();
+      await expect(navbar.getByRole('link', { name: /Releases/i })).toBeVisible();
+      await expect(navbar.getByRole('link', { name: /Deployments/i })).toBeVisible();
+    });
+    test('has clickable header buttons', async () => {
+      await expect(navbar.getByRole('link', { name: /Dashboard/i })).toBeVisible();
+      await navbar.getByRole('link', { name: /Dashboard/i }).click();
+      await navbar.getByRole('link', { name: /Devices/i }).click();
+      await navbar.getByRole('link', { name: /Releases/i }).click();
+      await navbar.getByRole('link', { name: /Deployments/i }).click();
     });
   });
 });
